Crate massa_models

Copyright (c) 2022 MASSA LABS info@massa.net All the structures that are used everywhere

Modules

active_block

active blocks related structures

address

address related structures

amount

amount related structures

async_msg

async message related structures Copyright (c) 2022 MASSA LABS info@massa.net This file defines the structure representing an asynchronous message

async_msg_id

async message id

block

block structure Copyright (c) 2022 MASSA LABS info@massa.net

block_header

block-related structure: block_header

block_id

block-related structure: block_id

bytecode

bytecode structures

clique

clique

composite

various structures

config

node configuration Main node configuration and all that stuff is here

datastore

datastore serialization / deserialization

deferred_calls

deferred call id

denunciation

denunciation Copyright (c) 2022 MASSA LABS info@massa.net Denunciation intro

endorsement

endorsements

error

models error

execution

execution related structures

ledger

ledger related structures

mapping_grpc

mapping grpc

node

node related structure

operation

operations

output_event

smart contract output events

prehash

pre-hashed trait, for hash less hashmap/set

rolls

rolls

secure_share

trait for massa_signature::Signature secured data-structs

serialization

serialization

slot

slots

stats

various statistics Copyright (c) 2022 MASSA LABS info@massa.net

streaming_step

bootstrap streaming cursor

timeslots

management of the relation between time and slots warning: assumes thread_count >= 1, t0_millis >= 1, t0_millis % thread_count == 0

types

types

version

versions