massa_models

Module address

address related structures

Macros

SCAddress 🔒

SCAddressVariant 🔒

UserAddress 🔒

UserAddressVariant 🔒

Structs

AddressDeserializer

Deserializer for Address

AddressSerializer

Serializer for Address

ExecutionAddressCycleInfo

Info for a given address on a given cycle

SCAddressV0

Derived from a public key.

UserAddressV0

Derived from a public key.

Enums

Address

Top level address representation that can differentiate between User and SC address

SCAddress

Derived from a public key.

UserAddress

Derived from a public key.

Constants

ADDRESS_PREFIX 🔒

SC_PREFIX 🔒

USER_PREFIX 🔒