massa_models::amount

Struct Amount

pub struct Amount(u64);

A structure representing a decimal Amount of coins with safe operations this allows ensuring that there is never an uncontrolled overflow or precision loss while providing a convenient decimal interface for users The underlying u64 raw representation if a fixed-point value with factor AMOUNT_DECIMAL_FACTOR The minimal value is 0 and the maximal value is 18446744073.709551615(std::u64::MAX/1e9)

Tuple Fields
0: u64
Implementations

impl Amount

pub const MIN: Amount

Minimum amount

pub const MAX: Amount

Maximum amount

pub const fn zero() -> Self

Create a zero Amount

fn to_decimal(self) -> Decimal

Convert to decimal

fn from_decimal(dec: Decimal) -> Result<Self, ModelsError>

Create an Amount from a Decimal

pub const fn const_init(mantissa: u64, scale: u32) -> Self

Create an Amount from the form mantissa / (10^scale) in a const way. WARNING: Panics on any error. Used only for constant initialization.

let amount_1: Amount = Amount::from_str("0.042").unwrap();
let amount_2: Amount = Amount::const_init(42, 3);
assert_eq!(amount_1, amount_2);
let amount_1: Amount = Amount::from_str("1000").unwrap();
let amount_2: Amount = Amount::const_init(1000, 0);
assert_eq!(amount_1, amount_2);

pub fn to_mantissa_scale(&self) -> (u64, u32)

Returns the value in the (mantissa, scale) format where amount = mantissa * 10^(-scale)

let amount = Amount::from_str("0.123456789").unwrap();
let (mantissa, scale) = amount.to_mantissa_scale();
assert_eq!(mantissa, 123456789);
assert_eq!(scale, AMOUNT_DECIMAL_SCALE);

pub fn from_mantissa_scale( mantissa: u64, scale: u32, ) -> Result<Self, ModelsError>

Creates an amount in the format mantissa*10^(-scale).

let a = Amount::from_mantissa_scale(123, 2).unwrap();
assert_eq!(a.to_string(), "1.23");
let a = Amount::from_mantissa_scale(123, 100);
assert!(a.is_err());

pub const fn to_raw(&self) -> u64

Obtains the underlying raw u64 representation Warning: do not use this unless you know what you are doing because the raw value does not take the AMOUNT_DECIMAL_FACTOR into account.

pub const fn from_raw(raw: u64) -> Self

constructs an Amount from the underlying raw u64 representation Warning: do not use this unless you know what you are doing because the raw value does not take the AMOUNT_DECIMAL_FACTOR into account In most cases, you should be using Amount::from_str("11.23")

pub fn saturating_add(self, amount: Amount) -> Self

safely add self to another amount, saturating the result on overflow

pub fn saturating_sub(self, amount: Amount) -> Self

safely subtract another amount from self, saturating the result on underflow

pub fn is_zero(&self) -> bool

returns true if the amount is zero

pub fn checked_sub(self, amount: Amount) -> Option<Self>

safely subtract another amount from self, returning None on underflow

let amount_1 : Amount = Amount::from_str("42").unwrap();
let amount_2 : Amount = Amount::from_str("7").unwrap();
let res : Amount = amount_1.checked_sub(amount_2).unwrap();
assert_eq!(res, Amount::from_str("35").unwrap())

pub fn checked_add(self, amount: Amount) -> Option<Self>

safely add self to another amount, returning None on overflow

let amount_1 : Amount = Amount::from_str("42").unwrap();
let amount_2 : Amount = Amount::from_str("7").unwrap();
let res : Amount = amount_1.checked_add(amount_2).unwrap();
assert_eq!(res, Amount::from_str("49").unwrap())

pub fn checked_mul_u64(self, factor: u64) -> Option<Self>

safely multiply self with a u64, returning None on overflow

let amount_1 : Amount = Amount::from_str("42").unwrap();
let res : Amount = amount_1.checked_mul_u64(7).unwrap();
assert_eq!(res, Amount::from_str("294").unwrap())

pub const fn saturating_mul_u64(self, factor: u64) -> Self

safely multiply self with a u64, saturating the result on overflow

let amount_1 : Amount = Amount::from_str("42").unwrap();
let res : Amount = amount_1.saturating_mul_u64(7);
assert_eq!(res, Amount::from_str("294").unwrap());

pub fn checked_div_u64(self, factor: u64) -> Option<Self>

safely divide self by a u64, returning None if the factor is zero

let amount_1 : Amount = Amount::from_str("42").unwrap();
let res : Amount = amount_1.checked_div_u64(7).unwrap();
assert_eq!(res, Amount::from_str("6").unwrap());

pub fn checked_div(self, divisor: Self) -> Option<u64>

safely divide self by an amount, returning None if the divisor is zero

let amount_1 : Amount = Amount::from_str("42").unwrap();
let amount_2 : Amount = Amount::from_str("7").unwrap();
let res : u64 = amount_1.checked_div(amount_2).unwrap();
assert_eq!(res, 6);

pub fn checked_rem(&self, divisor: &Amount) -> Option<Amount>

compute self % divisor, return None if divisor is zero

let amount_1 : Amount = Amount::from_str("42").unwrap();
let amount_2 : Amount = Amount::from_str("10").unwrap();
let res : Amount = amount_1.checked_rem(&amount_2).unwrap();
assert_eq!(res, Amount::from_str("2").unwrap());

pub fn checked_rem_u64(&self, divisor: u64) -> Option<Amount>

compute self % divisor, return None if divisor is zero

let amount_1 : Amount = Amount::from_str("42").unwrap();
let res : Amount = amount_1.checked_rem_u64(40000000000).unwrap();
assert_eq!(res, Amount::from_str("2").unwrap());