Module massa_models::constants

Expose constants

Modules

default

DEFAULT VALUES USED TO INITIALISE DIVERS CONFIGURATIONS STRUCTURES

default_testing

DEFAULT VALUES USED TO INITIALISE DIVERS CONFIGURATIONS STRUCTURES Same as default constants but in testing mode. You can access to them with the testing feature activated.

Structs

CompactConfig

Compact representation of key values of consensus algorithm used in API

END_TIMESTAMP

TESTNET: time when the blockclique is ended.

GENESIS_KEY

Private_key to sign genesis blocks.

GENESIS_TIMESTAMP

Time in millis when the blockclique started.

POS_MISS_RATE_DEACTIVATION_THRESHOLD

number of cycle misses (strictly) above which stakers are deactivated

VERSION

Constants

ADDRESS_SIZE_BYTES

AMOUNT_DECIMAL_FACTOR

Safe to import

BLOCK_ID_SIZE_BYTES

BLOCK_REWARD

Block reward is given for each block creation

BOOTSTRAP_RANDOMNESS_SIZE_BYTES

Size of the random bytes array used for the bootstrap, safe to import

CHANNEL_SIZE

DELTA_F0

Threshold for fitness.

ENDORSEMENT_COUNT

Number of endorsement

ENDORSEMENT_ID_SIZE_BYTES

EVENT_ID_SIZE_BYTES

Size of the event id hash used in execution module, safe to import

HANDSHAKE_RANDOMNESS_SIZE_BYTES

Length of the handshake random signature

INITIAL_DRAW_SEED

Proof of stake seed for the initial draw

IP_LIST_MAX_SIZE

Max size of the IP list

MAX_ADVERTISE_LENGTH

Limit on the number of peers we advertise to others.

MAX_ASK_BLOCKS_PER_MESSAGE

Max number of hash in the message AskForBlocks

MAX_BLOCK_SIZE

Maximum block size in bytes

MAX_BOOTSTRAP_BLOCKS

Max number of blocks we provide/ take into account while bootstrapping

MAX_BOOTSTRAP_CHILDREN

Max number of child nodes

MAX_BOOTSTRAP_CLIQUES

MAX_BOOTSTRAP_DEPS

MAX_BOOTSTRAP_MESSAGE_SIZE

Max message size for bootstrap

MAX_BOOTSTRAP_POS_CYCLES

Max number of cycles in PoS bootstrap

MAX_BOOTSTRAP_POS_ENTRIES

Max number of address and rng entries for PoS bootstrap

MAX_DUPLEX_BUFFER_SIZE

MAX_ENDORSEMENTS_PER_MESSAGE

Max number of endorsements per message

MAX_GAS_PER_BLOCK

Maximum of GAS allowed for a block

MAX_MESSAGE_SIZE

Maximum message length in bytes

MAX_OPERATIONS_PER_BLOCK

Maximum number of operations per block

MAX_OPERATIONS_PER_MESSAGE

Max number of operations per message

NODE_SEND_CHANNEL_SIZE

OPERATION_ID_SIZE_BYTES

OPERATION_VALIDITY_PERIODS

Maximum operation validity period count

PERIODS_PER_CYCLE

cycle duration in periods

POS_LOCK_CYCLES

PoS lock cycles: when some rolls are released, we only credit the coins back to their owner after waiting pos_lock_cycles

POS_LOOKBACK_CYCLES

PoS lookback cycles: when drawing for cycle N, we use the rolls from cycle N - pos_lookback_cycles - 1

ROLL_PRICE

Price of a roll in the network

SLOT_KEY_SIZE

T0

Time between the periods in the same thread.

THREAD_COUNT

Number of threads

Functions

build_massa_settings