Module massa_models::constants::default_testing

DEFAULT VALUES USED TO INITIALISE DIVERS CONFIGURATIONS STRUCTURES Same as default constants but in testing mode. You can access to them with the testing feature activated.

Default hardcoded

Each crates may contains a settings.rs or a config.rs the Default implementation of each object take the default Values from the following file.

These values are the hardcoded values that make sens to never be modifyed by a user. Generally, this values are passed with dependency injection in a cfg parameter for each worker, that is conveniant for unit tests.

See ./default.rs for more documentation about each constants

/!\ The following values are good as it. If you want a specific configuration, create a mutable configuration and inject it as a dependency in the worker.

Ex:

let cfg = UsedConfig {
   value_to_modify: 5,
   ..UsedConfig::default()
};

Structs

BLOCK_DB_PRUNE_INTERVAL

END_TIMESTAMP

TESTNET: time when the blockclique is ended.

GENESIS_KEY

GENESIS_TIMESTAMP

Time in millis when the blockclique started.

LEDGER_FLUSH_INTERVAL

MAX_SEND_WAIT

POS_MISS_RATE_DEACTIVATION_THRESHOLD

STATS_TIMESPAN

VERSION

Constants

ADDRESS_SIZE_BYTES

Size of the random bytes array used for the bootstrap, safe to import

AMOUNT_DECIMAL_FACTOR

Safe to import

BASE_NETWORK_CONTROLLER_IP

BLOCK_ID_SIZE_BYTES

BLOCK_REWARD

BOOTSTRAP_RANDOMNESS_SIZE_BYTES

CHANNEL_SIZE

CURSOR_DELAY

DELTA_F0

DISABLE_BLOCK_CREATION

ENDORSEMENT_COUNT

ENDORSEMENT_ID_SIZE_BYTES

EVENT_ID_SIZE_BYTES

FINAL_HISTORY_LENGTH

FORCE_KEEP_FINAL_PERIOD

FUTURE_BLOCK_PROCESSING_MAX_PERIODS

HANDSHAKE_RANDOMNESS_SIZE_BYTES

INITIAL_DRAW_SEED

IP_LIST_MAX_SIZE

LEDGER_CACHE_CAPACITY

LEDGER_RESET_AT_STARTUP

MAX_ADVERTISE_LENGTH

MAX_ASK_BLOCKS_PER_MESSAGE

MAX_BLOCK_SIZE

MAX_BOOTSTRAP_BLOCKS

MAX_BOOTSTRAP_CHILDREN

MAX_BOOTSTRAP_CLIQUES

MAX_BOOTSTRAP_DEPS

MAX_BOOTSTRAP_MESSAGE_SIZE

MAX_BOOTSTRAP_POS_CYCLES

MAX_BOOTSTRAP_POS_ENTRIES

MAX_DEPENDENCY_BLOCK

MAX_DISCARED_BLOCKS

MAX_DUPLEX_BUFFER_SIZE

MAX_ENDORSEMENTS_PER_MESSAGE

MAX_FINAL_EVENTS

MAX_FUTURE_PROCESSING_BLOCK

MAX_GAS_PER_BLOCK

MAX_ITEM_RETURN_COUNT

MAX_MESSAGE_SIZE

MAX_OPERATIONS_PER_BLOCK

MAX_OPERATIONS_PER_MESSAGE

MAX_OPERATION_FILL_ATTEMPTS

NODE_SEND_CHANNEL_SIZE

OPERATION_BATCH_SIZE

OPERATION_ID_SIZE_BYTES

OPERATION_VALIDITY_PERIODS

PERIODS_PER_CYCLE

POS_DRAW_CACHED_CYCLE

POS_LOCK_CYCLES

POS_LOOKBACK_CYCLES

READONLY_QUEUE_LENGTH

ROLL_PRICE

SLOT_KEY_SIZE

T0

THREAD_COUNT