Module massa_models::datastore

datastore serialization / deserialization

Structs

DatastoreDeserializer

Deserializer for Datastore field in LedgerEntry

DatastoreSerializer

Serializer for Datastore

Functions

get_prefix_bounds

For lexicographically ordered keys, gets the upper and lower bound of keys matching a prefix.

Type Definitions

Datastore

Datastore entry for Ledger & ExecuteSC Operation A Datastore is a Key Value store where Key: Byte array (max length should be 255) Value: Byte array What is stored can be arbitrary bytes but can often be smart contract bytecode (aka WASM binary)