Enum massa_models::denunciation::Denunciation

pub enum Denunciation {
    Endorsement(EndorsementDenunciation),
    BlockHeader(BlockHeaderDenunciation),
}

A denunciation enum

Variants

Endorsement(EndorsementDenunciation)

BlockHeader(BlockHeaderDenunciation)

Implementations

impl Denunciation

pub fn is_for_endorsement(&self) -> bool

Check if it is a Denunciation of several endorsements

pub fn is_for_block_header(&self) -> bool

Check if it is a Denunciation of several block headers

pub fn is_also_for_endorsement( &self, s_endorsement: &SecureShareEndorsement ) -> Result<bool, DenunciationError>

Check if it is a Denunciation for this endorsement

pub fn is_also_for_block_header( &self, s_block_header: &SecuredHeader ) -> Result<bool, DenunciationError>

Check if it is a Denunciation for this block header

pub fn is_valid(&self) -> bool

Check if Denunciation is valid Should be used if received from the network (prevent against invalid or attacker crafted denunciation)

pub fn get_slot(&self) -> &Slot

Get Denunciation slot ref

pub fn get_index(&self) -> Option<&u32>

Get field: index (return None for a block header denunciation)

pub fn get_public_key(&self) -> &PublicKey

Get Denunciation public key ref

pub fn is_expired( denunciation_slot_period: &u64, slot_period: &u64, denunciation_expire_periods: &u64 ) -> bool

Check if denunciation has expired given a slot period Note that slot_period can be:

A final slot period (for example in order to cleanup denunciation pool caches)

A block slot period (in execution (execute_denunciation(…)))