Module massa_models::endorsement

endorsements

Macros

EndorsementId 🔒

EndorsementIdVariant 🔒

Structs

Endorsement

an endorsement, as sent in the network

EndorsementDenunciationData

A denunciation data for endorsement

EndorsementDeserializer

Deserializer for Endorsement

EndorsementDeserializerLW

Lightweight Deserializer for Endorsement

EndorsementIdDeserializer 🔒

EndorsementIdV0

endorsement id

EndorsementSerializer

Serializer for Endorsement

EndorsementSerializerLW

Lightweight Serializer for Endorsement When included in a BlockHeader, we want to serialize only the index (optimization)

Enums

EndorsementId

endorsement id

Constants

ENDORSEMENTID_PREFIX 🔒

ENDORSEMENTID_VERSION 🔒

ENDORSEMENT_ID_SIZE_BYTES

Endorsement ID size in bytes

Type Aliases

SecureShareEndorsement

Wrapped endorsement