Struct massa_models::rolls::RollUpdates

pub struct RollUpdates(pub PreHashMap<Address, RollUpdate>);

maps addresses to roll updates

Tuple Fields
0: PreHashMap<Address, RollUpdate>
Implementations

impl RollUpdates

pub fn get_involved_addresses(&self) -> PreHashSet<Address>

the addresses impacted by the updates

pub fn chain(
&mut self,
updates: &RollUpdates
) -> Result<PreHashMap<Address, RollCompensation>>

chains with another RollUpdates, compensates and returns compensations

pub fn apply(
&mut self,
addr: &Address,
update: &RollUpdate
) -> Result<RollCompensation>

applies a RollUpdate, compensates and returns compensation

pub fn clone_subset(&self, addrs: &PreHashSet<Address>) -> Self

get the roll update for a subset of addresses

pub fn sync_from(&mut self, addrs: &PreHashSet<Address>, other: RollUpdates)

merge another roll updates into self, overwriting existing data addresses that are in not other are removed from self