massa_models

Module secure_share

trait for massa_signature::Signature secured data-structs

Structs

SecureShare

Packages type T such that it can be securely sent and received in a trust-free network

SecureShareDeserializer

Deserializer for SecureShare structure

SecureShareSerializer

Serializer for SecureShare structure

Traits

Id

Used by signed structure TODO: Make this trait use versions

SecureShareContent

Trait that define a structure that can be signed for secure sharing.