Struct massa_models::secure_share::SecureShare

pub struct SecureShare<T, ID>where
    T: Display + SecureShareContent,
    ID: Id,{
    pub content: T,
    pub serialized_data: Vec<u8>,
    pub signature: Signature,
    pub content_creator_pub_key: PublicKey,
    pub content_creator_address: Address,
    pub id: ID,
}

Packages type T such that it can be securely sent and received in a trust-free network

If the internal content is mutated, then it must be re-wrapped, as the assosciated signature, serialized data, etc. would no longer be in sync

Fields
content: T

Reference contents. Not required for the the security protocols.

Use the Lightweight equivilant structures when you need verifiable serialized data, but do not need to read the values directly (such as when sending)

serialized_data: Vec<u8>

Content in sharable, deserializable form. Is used in the secure verification protocols.

signature: Signature

A cryptographically generated value using serialized_data and a public key.

content_creator_pub_key: PublicKey

The public-key component used in the generation of the signature

content_creator_address: Address

Derived from the same public key used to generate the signature

id: ID

A secure hash of the data. See also [massa_hash::Hash]

Implementations

impl SecureShare<Block, BlockId>

pub fn bytes_count(&self) -> u64

size in bytes of the whole block

pub fn contains_operation(&self, op: SecureShareOperation) -> bool

true if given operation is included in the block

pub fn get_fitness(&self) -> u64

returns the fitness of the block

impl SecureShare<BlockHeader, BlockId>

pub fn get_fitness(&self) -> u64

gets the header fitness

impl SecureShare<Operation, OperationId>

pub fn get_validity_range( &self, operation_validity_period: u64 ) -> RangeInclusive<u64>

get the range of periods during which an operation is valid Range: (op.expire_period - cfg.operation_validity_period) -> op.expire_period (included)

pub fn get_gas_usage(&self) -> u64

Get the max amount of gas used by the operation (max_gas)

pub fn get_ledger_involved_addresses(&self) -> PreHashSet<Address>

get the addresses that are involved in this operation from a ledger point of view

pub fn get_max_spending(&self, roll_price: Amount) -> Amount

Gets the maximal amount of coins that may be spent by this operation (incl. fee)

pub fn get_roll_involved_addresses( &self ) -> Result<PreHashSet<Address>, ModelsError>

get the addresses that are involved in this operation from a rolls point of view

impl<T, ID> SecureShare<T, ID>where T: Display + SecureShareContent, ID: Id,

pub fn verify_signature(&self) -> Result<(), ModelsError>

check if self has been signed by public key

pub fn serialized_size(&self) -> usize

get full serialized size