massa_models::serialization

Trait DeserializeMinBEInt

pub trait DeserializeMinBEInt: Sized {
    // Required method
    fn from_be_bytes_min(
        buffer: &[u8],
        max_value: Self,
    ) -> Result<(Self, usize), ModelsError>;
}

Deserialize min big endian

Required Methods

fn from_be_bytes_min( buffer: &[u8], max_value: Self, ) -> Result<(Self, usize), ModelsError>

Deserializes a minimally sized big endian integer to Self from the provided buffer and checks that its value is within given bounds. In case of success, return the deserialized data and the number of bytes read

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.