Trait massa_models::signed::Signable

pub trait Signable<U> where
    U: Id,
    Self: SerializeCompact, {
    fn get_signature_message(&self) -> Result<Hash, ModelsError> { ... }
    fn compute_id(&self) -> Result<U, ModelsError> { ... }
}

implement if you want that structure to be signed

Provided Methods

fn get_signature_message(&self) -> Result<Hash, ModelsError>

The hash that should be used for the signature

fn compute_id(&self) -> Result<U, ModelsError>

Associated id