Trait massa_models::DeserializeVarInt

pub trait DeserializeVarInt: Sized {
    fn from_varint_bytes(buffer: &[u8]) -> Result<(Self, usize), ModelsError>;

    fn from_varint_bytes_bounded(
        buffer: &[u8], 
        max_value: Self
    ) -> Result<(Self, usize), ModelsError>;
}

Required methods

fn from_varint_bytes(buffer: &[u8]) -> Result<(Self, usize), ModelsError>

Deserialize variable size integer to Self from the provided buffer. The data to deserialize starts at the beginning of the buffer but the buffer can be larger than needed. In case of success, return the deserialized data and the number of bytes read

fn from_varint_bytes_bounded(
buffer: &[u8],
max_value: Self
) -> Result<(Self, usize), ModelsError>

Deserialize variable size integer to Self from the provided buffer and checks that its value is within given bounds. The data to deserialize starts at the beginning of the buffer but the buffer can be larger than needed. In case of success, return the deserialized data and the number of bytes read