Type Definition massa_models::WrappedBlock

pub type WrappedBlock = Wrapped<Block, BlockId>;

Wrapped Block

Implementations

impl WrappedBlock

pub fn bytes_count(&self) -> u64

size in bytes of the whole block

pub fn contains_operation(
&self,
op: WrappedOperation
) -> Result<bool, ModelsError>

true if given operation is included in the block may fail if computing an id of an operation in the block

pub fn get_roll_involved_addresses(&self) -> Result<Set<Address>, ModelsError>

Retrieve roll involving addresses

pub fn involved_addresses(
&self,
operation_set: &Map<OperationId, (usize, u64)>
) -> Result<Map<Address, Set<OperationId>>, ModelsError>

retrieves a mapping of addresses to the list of operation IDs they are involved with in terms of ledger

pub fn addresses_to_endorsements(
&self
) -> Result<Map<Address, Set<EndorsementId>>, ModelsError>

returns the set of addresses mapped the the endorsements they are involved in