Struct massa_models::wrapped::Wrapped

pub struct Wrapped<T, U>where
    T: Display + WrappedContent,
    U: Id,{
    pub content: T,
    pub signature: Signature,
    pub creator_public_key: PublicKey,
    pub creator_address: Address,
    pub id: U,
    pub serialized_data: Vec<u8>,
}

Wrapped structure T where U is the associated id

Fields
content: T

content

signature: Signature

signature

creator_public_key: PublicKey

the content creator public key

creator_address: Address

the content creator address

id: U

Id

serialized_data: Vec<u8>

Content serialized

Implementations

impl Wrapped<Block, BlockId>

pub fn bytes_count(&self) -> u64

size in bytes of the whole block

pub fn contains_operation(&self, op: WrappedOperation) -> bool

true if given operation is included in the block

pub fn get_fitness(&self) -> u64

returns the fitness of the block

impl Wrapped<BlockHeader, BlockId>

pub fn get_fitness(&self) -> u64

gets the header fitness

impl Wrapped<Operation, OperationId>

pub fn get_validity_range(
&self,
operation_validity_period: u64
) -> RangeInclusive<u64>

get the range of periods during which an operation is valid Range: (op.expire_period - cfg.operation_validity_period) -> op.expire_period (included)

pub fn get_gas_usage(&self) -> u64

Get the max amount of gas used by the operation (max_gas)

pub fn get_gas_price(&self) -> Amount

Get the gas price set by the operation

pub fn get_gas_coins(&self) -> Amount

Get the amount of coins used by the operation to pay for gas

pub fn get_total_fee(&self) -> Amount

Get the total fee paid by the creator

pub fn get_ledger_involved_addresses(&self) -> PreHashSet<Address>

get the addresses that are involved in this operation from a ledger point of view

pub fn get_max_spending(&self, roll_price: Amount) -> Amount

Gets the maximal amount of coins that may be spent by this operation (incl. fee)

pub fn get_roll_involved_addresses(
&self
) -> Result<PreHashSet<Address>, ModelsError>

get the addresses that are involved in this operation from a rolls point of view

impl<T, U> Wrapped<T, U>where
T: Display + WrappedContent,
U: Id,

pub fn verify_signature(&self) -> Result<(), ModelsError>

check if self has been signed by public key

pub fn serialized_size(&self) -> usize

get full serialized size