Crate massa_models

Copyright (c) 2022 MASSA LABS info@massa.net All the structures that are used everywhere

Modules

active_block

active blocks related structures

address

address related structures

amount

amount related structures

api

structure use by the API

block

block-related sturctures Copyright (c) 2022 MASSA LABS info@massa.net

clique

clique

composite

various structures

config

node config Main node configuration and all that stuff is here

endorsement

endorsements

error

models error

execution

execution related structures

ledger_models

ledger related structures

node

node related structure

operation

operations

output_event

smart contract output events

prehash

pre-hashed trait, for hash less hashmap/set

rolls

rolls

serialization

serialization

slot

slots

stats

various statistics Copyright (c) 2022 MASSA LABS info@massa.net

timeslots

management of the relation between time and slots warning: assumes thread_count >= 1, t0_millis >= 1, t0_millis % thread_count == 0

version

versions

wrapped

trait for signed structure