Crate massa_models

Re-exports

pub use address::Address;

pub use amount::Amount;

pub use composite::OperationSearchResult;

pub use composite::OperationSearchResultBlockStatus;

pub use composite::OperationSearchResultStatus;

pub use composite::StakersCycleProductionStats;

pub use error::ModelsError;

pub use operation::Operation;

pub use operation::OperationId;

pub use operation::OperationType;

pub use operation::SignedOperation;

pub use slot::Slot;

Modules

active_block

address

amount

api

clique

composite

constants

Expose constants

error

execution

ledger_models

node

operation

output_event

prehash

rolls

signed

slot

stats

timeslots

warning: assumes thread_count >= 1, t0_millis >= 1, t0_millis % thread_count == 0

Structs

Block

BlockHeader

BlockId

CompactConfig

Compact representation of key values of consensus algorithm used in API

Endorsement

EndorsementId

SerializationContext

a context for model serialization/deserialization

Version

Application version, checked during handshakes

Traits

DeserializeCompact

DeserializeMinBEInt

DeserializeVarInt

SerializeCompact

SerializeMinBEInt

SerializeVarInt

Functions

array_from_slice

get_serialization_context

Get a clone of the context. For tests only.

init_serialization_context

Initialize the global context, should be called once at startup or used at the beginning of a test.

u8_from_slice

with_serialization_context

Use the tls context, should be called only after initializing the global context.

Type Definitions

SignedEndorsement

SignedHeader