Crate massa_models

Re-exports

pub use address::Address;

pub use amount::Amount;

pub use error::ModelsError;

pub use operation::Operation;

pub use operation::OperationContent;

pub use operation::OperationHashMap;

pub use operation::OperationHashSet;

pub use operation::OperationId;

pub use operation::OperationType;

pub use slot::Slot;

Modules

address

amount

api

clique

error

hhasher

ledger

massa_hash

node

operation

slot

stats

timeslots

warning: assumes thread_count >= 1, t0_millis >= 1, t0_millis % thread_count == 0

Structs

Block

BlockHeader

BlockHeaderContent

BlockId

CompactConfig

Compact representation of key values of consensus algorithm used in API

Endorsement

EndorsementContent

EndorsementId

OperationSearchResult

SerializationContext

a context for model serialization/deserialization

StakersCycleProductionStats

Version

Application version, checked during handshakes

Enums

OperationSearchResultBlockStatus

OperationSearchResultStatus

Constants

ADDRESS_SIZE_BYTES

BLOCK_ID_SIZE_BYTES

OPERATION_ID_SIZE_BYTES

SLOT_KEY_SIZE

Traits

DeserializeCompact

DeserializeMinBEInt

DeserializeVarInt

SerializeCompact

SerializeMinBEInt

SerializeVarInt

Functions

array_from_slice

get_serialization_context

Get a clone of the context. For tests only.

init_serialization_context

Initialize the global context, should be called once at startup or used at the beginning of a test.

u8_from_slice

with_serialization_context

Use the tls context, should be called only after initializing the global context.

Type Definitions

BlockHashMap

BlockHashSet

EndorsementHashMap

EndorsementHashSet