Module massa_models::api

structure use by the API

Structs

AddressInfo

All you ever dream to know about an address

BlockInfo

refactor to delete

BlockInfoContent

Block content

BlockSummary

A block resume (without the block itself)

CompactAddressInfo

Less information about an address

DatastoreEntryInput

Datastore entry query input structure

DatastoreEntryOutput

Datastore entry query output structure

EndorsementInfo

All you wanna know about an endorsement

EventFilter

filter used when retrieving SC output events

IndexedSlot

When an address is drawn to create an endorsement it is selected for a specific index

LedgerInfo

Current balance ledger info

NodeStatus

node status

OperationInfo

Operation and contextual info about it

OperationInput

operation input

ReadOnlyBytecodeExecution

read only bytecode execution request

ReadOnlyCall

read SC call request

RollsInfo

Roll counts

SlotAmount

slot / amount pair

TimeInterval

Just a wrapper with a optional beginning and end

Enums

BlockGraphStatus

Block status within the graph