Module massa_models::api

structure use by the API

Structs

AddressInfo

All you ever dream to know about an address

BlockInfo

refactor to delete

BlockInfoContent

Block content

BlockSummary

A block resume (without the block itself)

CompactAddressInfo

Less information about an address

DatastoreEntryInput

Datastore entry query input struct

DatastoreEntryOutput

Datastore entry query output struct

EndorsementInfo

All you wanna know about an endorsement

EventFilter

filter used when retrieving SC output events

IndexedSlot

When an address is drawn to create an endorsement it is selected for a specific index

LedgerInfo

Current Parallel balance ledger info

NodeStatus

node status

OperationInfo

Operation and contextual info about it

ReadOnlyBytecodeExecution

read only bytecode execution request

ReadOnlyCall

read SC call request

RollsInfo

Roll counts

TimeInterval

Just a wrapper with a optional beginning and end