Module massa_models::config::constants

DEFAULT VALUES USED TO INITIALIZE DIVERS CONFIGURATIONS STRUCTURES

Default hard-coded

Each crates may contains a settings.rs or a config.rs the Default implementation of each object take the default Values from the following file.

These values are the hard-coded values that make sens to never be modified by a user. Generally, this values are passed with dependency injection in a cfg parameter for each worker, that is convenient for unit tests.

A parallel file with the same constant definitions exist for the testing case. (default_testing.rs) But as for the current file you shouldn’t modify it.

Structs

END_TIMESTAMP

TESTNET: time when the blockclique is ended.

GENESIS_KEY

KeyPair to sign genesis blocks.

GENESIS_TIMESTAMP

Time in milliseconds when the blockclique started.

POS_MISS_RATE_DEACTIVATION_THRESHOLD

number of cycle misses (strictly) above which stakers are deactivated

VERSION

node version

Constants

ASYNC_POOL_PART_SIZE_MESSAGE_BYTES

Maximum async messages in a batch of the bootstrap of the async pool

BLOCK_REWARD

Block reward is given for each block creation

BOOTSTRAP_RANDOMNESS_SIZE_BYTES

Size of the random bytes array used for the bootstrap, safe to import

CHANNEL_SIZE

Consensus static parameters (defined by protocol used) Changing one of the following values is considered as a breaking change Values differ in test flavor building for faster CI and simpler scenarios

CYCLE_INFO_SIZE_MESSAGE_BYTES

Maximum cycle info in a bootstrap batch

DEFERRED_CREDITS_PART_SIZE_MESSAGE_BYTES

Maximum deferred credits in a bootstrap batch

DELTA_F0

Threshold for fitness.

ENDORSEMENT_COUNT

Number of endorsement

HANDSHAKE_RANDOMNESS_SIZE_BYTES

Length of the handshake random signature

INITIAL_DRAW_SEED

Proof of stake seed for the initial draw

IP_LIST_MAX_SIZE

Max size of the IP list

LEDGER_COST_PER_BYTE

Cost to store one byte in the ledger

LEDGER_ENTRY_BASE_SIZE

Cost for a base entry (address + balance (5 bytes constant))

LEDGER_ENTRY_DATASTORE_BASE_SIZE

Cost for a base entry datastore 10 bytes constant to avoid paying more for longer keys

LEDGER_PART_SIZE_MESSAGE_BYTES

Maximum size batch of data in a part of the ledger

MAX_ADVERTISE_LENGTH

Limit on the number of peers we advertise to others.

MAX_ASK_BLOCKS_PER_MESSAGE

Max number of hash in the message AskForBlocks

MAX_ASYNC_GAS

Maximum of GAS allowed for asynchronous messages execution on one slot

MAX_ASYNC_POOL_LENGTH

Maximum capacity of the asynchronous messages pool

MAX_BLOCK_SIZE

Maximum block size in bytes

MAX_BOOTSTRAP_ASYNC_POOL_CHANGES

Max async pool changes

MAX_BOOTSTRAP_BLOCKS

Max number of blocks we provide/ take into account while bootstrapping

MAX_BOOTSTRAP_CHILDREN

Max number of child nodes

MAX_BOOTSTRAP_CLIQUES

max bootstrapped cliques

MAX_BOOTSTRAP_DEPS

max bootstrapped dependencies

MAX_BOOTSTRAP_ERROR_LENGTH

Max size of the printed error

MAX_BOOTSTRAP_FINAL_STATE_PARTS_SIZE

Max bytes in final states parts

MAX_BOOTSTRAP_MESSAGE_SIZE

Max message size for bootstrap

MAX_BOOTSTRAP_POS_CYCLES

Max number of cycles in PoS bootstrap

MAX_BOOTSTRAP_POS_ENTRIES

Max number of address and random entries for PoS bootstrap

MAX_BYTECODE_LENGTH

Maximum length of a datastore value

MAX_DATASTORE_ENTRY_COUNT

Maximum number of key/values in the datastore of a ledger entry

MAX_DATASTORE_KEY_LENGTH

Maximum length of a datastore key

MAX_DATASTORE_VALUE_LENGTH

Maximum length of a datastore value

MAX_DATA_ASYNC_MESSAGE

Maximum data size in async message

MAX_DUPLEX_BUFFER_SIZE

max duplex buffer size

MAX_ENDORSEMENTS_PER_MESSAGE

Max number of endorsements per message

MAX_FUNCTION_NAME_LENGTH

Maximum length function name in call SC

MAX_GAS_PER_BLOCK

Maximum of GAS allowed for a block

MAX_LEDGER_CHANGES_COUNT

Maximum ledger changes count

MAX_LEDGER_CHANGES_PER_SLOT

Maximum ledger changes in a block

MAX_MESSAGE_SIZE

Maximum message length in bytes

MAX_OPERATIONS_PER_BLOCK

Maximum number of operations per block

MAX_OPERATIONS_PER_MESSAGE

Max number of operations per message

MAX_OPERATION_DATASTORE_ENTRY_COUNT

Maximum number of key/values in the datastore of a ExecuteSC operation

MAX_OPERATION_DATASTORE_KEY_LENGTH

Maximum length of an operation datastore key

MAX_OPERATION_DATASTORE_VALUE_LENGTH

Maximum length of an operation datastore value

MAX_PARAMETERS_SIZE

Maximum size of parameters in call SC

MAX_PRODUCTION_EVENTS_PER_BLOCK

Maximum production events in a block

MAX_PRODUCTION_STATS_LENGTH

Maximum length of production_stats in thread cycle

MAX_RNG_SEED_LENGTH

Maximum length of rng_seed in thread cycle

MAX_ROLLS_COUNTS_LENGTH

Maximum length of rolls_counts in thread cycle

MAX_ROLLS_UPDATE_LENGTH

Maximum length of rolls_update in thread cycle

NETWORK_CONTROLLER_CHANNEL_SIZE

network controller communication channel size

NETWORK_EVENT_CHANNEL_SIZE

network event channel size

NETWORK_NODE_COMMAND_CHANNEL_SIZE

network node command channel size

NETWORK_NODE_EVENT_CHANNEL_SIZE

network node event channel size

NODE_SEND_CHANNEL_SIZE

node send channel size

OPERATION_VALIDITY_PERIODS

Maximum operation validity period count

PERIODS_PER_CYCLE

cycle duration in periods

POOL_CONTROLLER_CHANNEL_SIZE

Pool controller channel size

POS_SAVED_CYCLES

PoS saved cycles: number of cycles saved in PoSFinalState

PRODUCTION_STATS_PART_SIZE_MESSAGE_BYTES

Maximum production stats in a bootstrap batch

PROTOCOL_CONTROLLER_CHANNEL_SIZE

Protocol controller channel size

PROTOCOL_EVENT_CHANNEL_SIZE

Protocol event channel size

ROLL_COUNTS_PART_SIZE_MESSAGE_BYTES

Maximum rolls in a bootstrap batch

ROLL_PRICE

Price of a roll in the network

T0

Time between the periods in the same thread.

THREAD_COUNT

Number of threads