Module massa_models::operation

operations

Macros

OperationId 🔒

OperationIdVariant 🔒

OperationPrefixId 🔒

OperationPrefixIdVariant 🔒

Structs

Operation

the operation as sent in the network

OperationDeserializer

Serializer for Operation

OperationIdDeserializer

Deserializer for OperationId

OperationIdSerializer

Serializer for OperationId

OperationIdV0

operation id

OperationIdsDeserializer

Deserializer for Vec<OperationId>

OperationIdsSerializer

Serializer for Vec<OperationId>

OperationPrefixIdDeserializer

Deserializer for [OperationPrefixId]

OperationPrefixIdV0

Left part of the operation id hash stored in a vector of size OPERATION_ID_PREFIX_SIZE_BYTES

OperationPrefixIdsDeserializer

Deserializer for OperationPrefixIds

OperationPrefixIdsSerializer

Serializer for OperationPrefixIds

OperationSerializer

Serializer for Operation

OperationTypeDeserializer

Deserializer for OperationType

OperationTypeSerializer

Serializer for OperationType

OperationsDeserializer

Deserializer for Operations

OperationsSerializer

Serializer for Operations

Enums

OperationId

operation id

OperationPrefixId

Left part of the operation id hash stored in a vector of size OPERATION_ID_PREFIX_SIZE_BYTES

OperationType

Type specific operation content

OperationTypeId 🔒

Constants

OPERATIONID_PREFIX 🔒

OPERATION_ID_PREFIX_SIZE_BYTES

Size in bytes of the serialized operation ID prefix

Functions

compute_operations_hash

Compute the hash of a list of operations(used typically in block headers)

Type Aliases

OperationPrefixIds

Set of operation id’s prefix

SecureShareOperation

signed operation