Type Definition massa_models::operation::SecureShareOperation

pub type SecureShareOperation = SecureShare<Operation, OperationId>;

signed operation

Implementations

impl SecureShareOperation

pub fn get_validity_range( &self, operation_validity_period: u64 ) -> RangeInclusive<u64>

get the range of periods during which an operation is valid Range: (op.expire_period - cfg.operation_validity_period) -> op.expire_period (included)

pub fn get_gas_usage(&self) -> u64

Get the max amount of gas used by the operation (max_gas)

pub fn get_ledger_involved_addresses(&self) -> PreHashSet<Address>

get the addresses that are involved in this operation from a ledger point of view

pub fn get_max_spending(&self, roll_price: Amount) -> Amount

Gets the maximal amount of coins that may be spent by this operation (incl. fee)

pub fn get_roll_involved_addresses( &self ) -> Result<PreHashSet<Address>, ModelsError>

get the addresses that are involved in this operation from a rolls point of view