Trait massa_models::secure_share::SecureShareContent

pub trait SecureShareContentwhere
    Self: Sized + Display,{
    fn new_verifiable<Ser: Serializer<Self>, ID: Id>(
        content: Self,
        content_serializer: Ser,
        keypair: &KeyPair
    ) -> Result<SecureShare<Self, ID>, ModelsError> { ... }
    fn serialize(
        signature: &Signature,
        creator_public_key: &PublicKey,
        serialized_content: &[u8],
        buffer: &mut Vec<u8>
    ) -> Result<(), SerializeError> { ... }
    fn deserialize<'a, E: ParseError<&'a [u8]> + ContextError<&'a [u8]>, Deser: Deserializer<Self>, ID: Id>(
        content_serializer: Option<&dyn Serializer<Self>>,
        signature_deserializer: &SignatureDeserializer,
        creator_public_key_deserializer: &PublicKeyDeserializer,
        content_deserializer: &Deser,
        buffer: &'a [u8]
    ) -> IResult<&'a [u8], SecureShare<Self, ID>, E> { ... }
}

Trait that define a structure that can be signed for secure sharing.

Provided Methods

fn new_verifiable<Ser: Serializer<Self>, ID: Id>(
content: Self,
content_serializer: Ser,
keypair: &KeyPair
) -> Result<SecureShare<Self, ID>, ModelsError>

Using the provided key-pair, applies a cryptographic signature, and packages the data required to share and verify the data in a trust-free network of peers.

fn serialize(
signature: &Signature,
creator_public_key: &PublicKey,
serialized_content: &[u8],
buffer: &mut Vec<u8>
) -> Result<(), SerializeError>

Serialize the secured structure

fn deserialize<'a, E: ParseError<&'a [u8]> + ContextError<&'a [u8]>, Deser: Deserializer<Self>, ID: Id>(
content_serializer: Option<&dyn Serializer<Self>>,
signature_deserializer: &SignatureDeserializer,
creator_public_key_deserializer: &PublicKeyDeserializer,
content_deserializer: &Deser,
buffer: &'a [u8]
) -> IResult<&'a [u8], SecureShare<Self, ID>, E>

Deserialize the secured structure