Module massa_models::serialization

serialization

Structs

BitVecDeserializer

BitVec<u8> Deserializer

BitVecSerializer

BitVec<u8> Serializer

IpAddrDeserializer

Deserializer for IpAddr

IpAddrSerializer

Serializer for IpAddr

PreHashSetDeserializer

Basic PreHashSet<_> deserializer

PreHashSetSerializer

Basic PreHashSet<_> serializer

StringDeserializer

Deserializer for String with generic deserializer for the size of the string

StringSerializer

Serializer for String with generic serializer for the size of the string

VecDeserializer

Basic Vec<_> deserializer

VecSerializer

Basic Vec<_> serializer

VecU8Deserializer

Basic Vec<u8> deserializer

VecU8Serializer

Basic Vec<u8> serializer

Traits

DeserializeMinBEInt

Deserialize min big endian

SerializeMinBEInt

Serialize min big endian integer

Functions

array_from_slice

array from slice

u8_from_slice

u8 from slice

u32_be_bytes_min_length

Compute the minimal big endian deserialization size

u64_be_bytes_min_length

Compute the minimal big endian deserialization size