Trait massa_models::serialization::DeserializeMinBEInt

pub trait DeserializeMinBEInt: Sized {
    const MIN_BE_INT_BASE_SIZE: usize;

    fn be_bytes_min_length(max_value: Self) -> usize;
    fn from_be_bytes_min(
        buffer: &[u8],
        max_value: Self
    ) -> Result<(Self, usize), ModelsError>;
}

Deserialize min big endian

Required Associated Constants

const MIN_BE_INT_BASE_SIZE: usize

min big endian integer base size

Required Methods

fn be_bytes_min_length(max_value: Self) -> usize

Compute the minimal big endian deserialization size

fn from_be_bytes_min(
buffer: &[u8],
max_value: Self
) -> Result<(Self, usize), ModelsError>

Deserializes a minimally sized big endian integer to Self from the provided buffer and checks that its value is within given bounds. In case of success, return the deserialized data and the number of bytes read