Struct massa_models::wrapped::Wrapped

pub struct Wrapped<T, U> where
    T: Display + WrappedContent,
    U: Id,  {
    pub content: T,
    pub signature: Signature,
    pub creator_public_key: PublicKey,
    pub creator_address: Address,
    pub thread: u8,
    pub id: U,
    pub serialized_data: Vec<u8>,
}

Wrapped structure T where U is the associated id

Fields
content: T

content

signature: Signature

signature

creator_public_key: PublicKey

the content creator public key

creator_address: Address

the content creator address

thread: u8

Thread of the operation creator

id: U

Id

serialized_data: Vec<u8>

Content serialized

Implementations

impl Wrapped<Block, BlockId>

pub fn bytes_count(&self) -> u64

size in bytes of the whole block

pub fn contains_operation(
&self,
op: WrappedOperation
) -> Result<bool, ModelsError>

true if given operation is included in the block may fail if computing an id of an operation in the block

pub fn get_roll_involved_addresses(&self) -> Result<Set<Address>, ModelsError>

Retrieve roll involving addresses

pub fn involved_addresses(
&self,
operation_set: &Map<OperationId, (usize, u64)>
) -> Result<Map<Address, Set<OperationId>>, ModelsError>

retrieves a mapping of addresses to the list of operation IDs they are involved with in terms of ledger

pub fn addresses_to_endorsements(
&self
) -> Result<Map<Address, Set<EndorsementId>>, ModelsError>

returns the set of addresses mapped the the endorsements they are involved in

impl Wrapped<Operation, OperationId>

pub fn verify_integrity(&self) -> Result<OperationId, ModelsError>

Verifies the signature and integrity of the operation and computes operation ID

impl Wrapped<Operation, OperationId>

pub fn get_validity_range(
&self,
operation_validity_period: u64
) -> RangeInclusive<u64>

get the range of periods during which an operation is valid

pub fn get_gas_usage(&self) -> u64

Get the amount of gas used by the operation

pub fn get_gas_coins(&self) -> Amount

Get the amount of coins used by the operation to pay for gas

pub fn get_ledger_involved_addresses(&self) -> Set<Address>

get the addresses that are involved in this operation from a ledger point of view

pub fn get_roll_involved_addresses(&self) -> Result<Set<Address>, ModelsError>

get the addresses that are involved in this operation from a rolls point of view

impl<T, U> Wrapped<T, U> where
T: Display + WrappedContent,
U: Id,

pub fn verify_signature<SC: Serializer<T>>(
&self,
content_serializer: SC,
public_key: &PublicKey
) -> Result<(), ModelsError>

check if self has been signed by public key